project/requirements.txt
jupyter
pandas
numpy
python-dotenv
mysql-connector-python
python-decouple
scikit-learn
matplotlib